requirements.txt
imageio>=2.8.0
matplotlib==3.2.1
opencv-python>=4.2.0.34
Pillow>=6.1.0
scipy>=1.4.1
